Add ID's to helper classes
