+++ WrathCombo/Combos/PvE/BRD/BRD.cs
@@ -11,75 +11,6 @@ namespace WrathCombo.Combos.PvE;
 
 internal partial class BRD
 {
-    public const byte ClassID = 5;
-    public const byte JobID = 23;
-
-    public const uint
-        HeavyShot = 97,
-        StraightShot = 98,
-        VenomousBite = 100,
-        RagingStrikes = 101,
-        QuickNock = 106,
-        Barrage = 107,
-        Bloodletter = 110,
-        Windbite = 113,
-        MagesBallad = 114,
-        ArmysPaeon = 116,
-        RainOfDeath = 117,
-        BattleVoice = 118,
-        EmpyrealArrow = 3558,
-        WanderersMinuet = 3559,
-        IronJaws = 3560,
-        TheWardensPaeon = 3561,
-        Sidewinder = 3562,
-        PitchPerfect = 7404,
-        Troubadour = 7405,
-        CausticBite = 7406,
-        Stormbite = 7407,
-        RefulgentArrow = 7409,
-        BurstShot = 16495,
-        ApexArrow = 16496,
-        Shadowbite = 16494,
-        Ladonsbite = 25783,
-        BlastArrow = 25784,
-        RadiantFinale = 25785,
-        WideVolley = 36974,
-        HeartbreakShot = 36975,
-        ResonantArrow = 36976,
-        RadiantEncore = 36977;
-
-    public static class Buffs
-    {
-        public const ushort
-            RagingStrikes = 125,
-            Barrage = 128,
-            MagesBallad = 135,
-            ArmysPaeon = 137,
-            BattleVoice = 141,
-            WanderersMinuet = 865,
-            Troubadour = 1934,
-            BlastArrowReady = 2692,
-            RadiantFinale = 2722,
-            ShadowbiteReady = 3002,
-            HawksEye = 3861,
-            ResonantArrowReady = 3862,
-            RadiantEncoreReady = 3863;
-    }
-
-    public static class Debuffs
-    {
-        public const ushort
-            VenomousBite = 124,
-            Windbite = 129,
-            CausticBite = 1200,
-            Stormbite = 1201;
-    }
-
-    internal static class Traits
-    {
-        internal const ushort
-            EnhancedBloodletter = 445;
-    }
     #region Song status
     internal static bool SongIsNotNone(Song value) => value != Song.NONE;
     internal static bool SongIsNone(Song value) => value == Song.NONE;
