+++ WrathCombo/Combos/PvE/DNC/DNC.cs
@@ -1502,94 +1502,4 @@ internal partial class DNC
     #endregion
 
     #endregion
-
-    #region IDs
-
-    public const byte JobID = 38;
-
-    #region Actions
-
-    public const uint
-        // Single Target
-        Cascade = 15989,
-        Fountain = 15990,
-        ReverseCascade = 15991,
-        Fountainfall = 15992,
-        StarfallDance = 25792,
-        // AoE
-        Windmill = 15993,
-        Bladeshower = 15994,
-        RisingWindmill = 15995,
-        Bloodshower = 15996,
-        Tillana = 25790,
-        // Dancing
-        StandardStep = 15997,
-        TechnicalStep = 15998,
-        StandardFinish0 = 16003,
-        StandardFinish1 = 16191,
-        StandardFinish2 = 16192,
-        TechnicalFinish0 = 16004,
-        TechnicalFinish1 = 16193,
-        TechnicalFinish2 = 16194,
-        TechnicalFinish3 = 16195,
-        TechnicalFinish4 = 16196,
-        Emboite = 15999,
-        Entrechat = 16000,
-        Jete = 16001,
-        Pirouette = 16002,
-        // Fan Dances
-        FanDance1 = 16007,
-        FanDance2 = 16008,
-        FanDance3 = 16009,
-        FanDance4 = 25791,
-        // Other
-        Peloton = 7557,
-        SaberDance = 16005,
-        ClosedPosition = 16006,
-        EnAvant = 16010,
-        Devilment = 16011,
-        ShieldSamba = 16012,
-        Flourish = 16013,
-        Improvisation = 16014,
-        CuringWaltz = 16015,
-        LastDance = 36983,
-        FinishingMove = 36984,
-        DanceOfTheDawn = 36985;
-
-    #endregion
-
-    public static class Buffs
-    {
-        public const ushort
-            // Flourishing & Silken (procs)
-            FlourishingCascade = 1814,
-            FlourishingFountain = 1815,
-            FlourishingWindmill = 1816,
-            FlourishingShower = 1817,
-            FlourishingFanDance = 2021,
-            SilkenSymmetry = 2693,
-            SilkenFlow = 2694,
-            FlourishingFinish = 2698,
-            FlourishingStarfall = 2700,
-            FlourishingSymmetry = 3017,
-            FlourishingFlow = 3018,
-            // Dances
-            StandardStep = 1818,
-            TechnicalStep = 1819,
-            StandardFinish = 1821,
-            TechnicalFinish = 1822,
-            // Fan Dances
-            ThreeFoldFanDance = 1820,
-            FourFoldFanDance = 2699,
-            // Other
-            Peloton = 1199,
-            ClosedPosition = 1823,
-            ShieldSamba = 1826,
-            LastDanceReady = 3867,
-            FinishingMoveReady = 3868,
-            DanceOfTheDawnReady = 3869,
-            Devilment = 1825;
-    }
-
-    #endregion
 }
