+++ WrathCombo/Combos/PvE/DRG/DRG.cs
@@ -12,7 +12,8 @@ internal partial class DRG
 
         protected override uint Invoke(uint actionID)
         {
-            if (actionID is not (FullThrust or HeavensThrust)) return actionID;
+            if (actionID is not (FullThrust or HeavensThrust))
+                return actionID;
 
             if (ComboTimer > 0)
             {
@@ -30,7 +31,6 @@ internal partial class DRG
             }
 
             return OriginalHook(TrueThrust);
-
         }
     }
 
@@ -40,7 +40,8 @@ internal partial class DRG
 
         protected override uint Invoke(uint actionID)
         {
-            if (actionID is not (ChaosThrust or ChaoticSpring)) return actionID;
+            if (actionID is not (ChaosThrust or ChaoticSpring))
+                return actionID;
 
             if (ComboTimer > 0)
             {
@@ -58,7 +59,6 @@ internal partial class DRG
             }
 
             return OriginalHook(TrueThrust);
-
         }
     }
 
@@ -723,74 +723,4 @@ internal partial class DRG
                 ? BattleLitany
                 : actionID;
     }
-
-    #region ID's
-
-    public const byte ClassID = 4;
-    public const byte JobID = 22;
-
-    public const uint
-        PiercingTalon = 90,
-        ElusiveJump = 94,
-        LanceCharge = 85,
-        BattleLitany = 3557,
-        Jump = 92,
-        LifeSurge = 83,
-        HighJump = 16478,
-        MirageDive = 7399,
-        BloodOfTheDragon = 3553,
-        Stardiver = 16480,
-        CoerthanTorment = 16477,
-        DoomSpike = 86,
-        SonicThrust = 7397,
-        ChaosThrust = 88,
-        RaidenThrust = 16479,
-        TrueThrust = 75,
-        Disembowel = 87,
-        FangAndClaw = 3554,
-        WheelingThrust = 3556,
-        FullThrust = 84,
-        VorpalThrust = 78,
-        WyrmwindThrust = 25773,
-        DraconianFury = 25770,
-        ChaoticSpring = 25772,
-        DragonfireDive = 96,
-        Geirskogul = 3555,
-        Nastrond = 7400,
-        HeavensThrust = 25771,
-        Drakesbane = 36952,
-        RiseOfTheDragon = 36953,
-        LanceBarrage = 36954,
-        SpiralBlow = 36955,
-        Starcross = 36956;
-
-    public static class Buffs
-    {
-        public const ushort
-            LanceCharge = 1864,
-            BattleLitany = 786,
-            DiveReady = 1243,
-            RaidenThrustReady = 1863,
-            PowerSurge = 2720,
-            LifeSurge = 116,
-            DraconianFire = 1863,
-            NastrondReady = 3844,
-            StarcrossReady = 3846,
-            DragonsFlight = 3845;
-    }
-
-    public static class Debuffs
-    {
-        public const ushort
-            ChaosThrust = 118,
-            ChaoticSpring = 2719;
-    }
-
-    public static class Traits
-    {
-        public const uint
-            EnhancedLifeSurge = 438;
-    }
-
-    #endregion
 }
