+++ WrathCombo/Combos/PvE/GNB/GNB.cs
@@ -14,91 +14,6 @@ namespace WrathCombo.Combos.PvE;
 
 internal partial class GNB
 {
-    public const byte JobID = 37; //Gunbreaker (GNB)
-
-    public const uint //Actions
-    #region Offensive
-        KeenEdge = 16137, //Lv1, instant, GCD, range 3, single-target, targets=hostile
-        NoMercy = 16138, //Lv2, instant, 60.0s CD (group 10), range 0, single-target, targets=self
-        BrutalShell = 16139, //Lv4, instant, GCD, range 3, single-target, targets=hostile
-        DemonSlice = 16141, //Lv10, instant, GCD, range 0, AOE 5 circle, targets=self
-        LightningShot = 16143, //Lv15, instant, GCD, range 20, single-target, targets=hostile
-        DangerZone = 16144, //Lv18, instant, 30s CD (group 4), range 3, single-target, targets=hostile
-        SolidBarrel = 16145, //Lv26, instant, GCD, range 3, single-target, targets=hostile
-        BurstStrike = 16162, //Lv30, instant, GCD, range 3, single-target, targets=hostile
-        DemonSlaughter = 16149, //Lv40, instant, GCD, range 0, AOE 5 circle, targets=self
-        SonicBreak = 16153, //Lv54, instant, 60.0s CD (group 13/57), range 3, single-target, targets=hostile
-        GnashingFang = 16146, //Lv60, instant, 30.0s CD (group 5/57), range 3, single-target, targets=hostile, animLock=0.700
-        SavageClaw = 16147, //Lv60, instant, GCD, range 3, single-target, targets=hostile, animLock=0.500
-        WickedTalon = 16150, //Lv60, instant, GCD, range 3, single-target, targets=hostile, animLock=0.770
-        BowShock = 16159, //Lv62, instant, 60.0s CD (group 11), range 0, AOE 5 circle, targets=self
-        AbdomenTear = 16157, //Lv70, instant, 1.0s CD (group 0), range 5, single-target, targets=hostile
-        JugularRip = 16156, //Lv70, instant, 1.0s CD (group 0), range 5, single-target, targets=hostile
-        EyeGouge = 16158, //Lv70, instant, 1.0s CD (group 0), range 5, single-target, targets=hostile
-        Continuation = 16155, //Lv70, instant, 1.0s CD (group 0), range 0, single-target, targets=self, animLock=???
-        FatedCircle = 16163, //Lv72, instant, GCD, range 0, AOE 5 circle, targets=self
-        Bloodfest = 16164, //Lv76, instant, 120.0s CD (group 14), range 25, single-target, targets=hostile
-        BlastingZone = 16165, //Lv80, instant, 30.0s CD (group 4), range 3, single-target, targets=hostile
-        Hypervelocity = 25759, //Lv86, instant, 1.0s CD (group 0), range 5, single-target, targets=hostile
-        DoubleDown = 25760, //Lv90, instant, 60.0s CD (group 12/57), range 0, AOE 5 circle, targets=self
-        FatedBrand = 36936, //Lv96, instant, 1.0s CD, (group 0), range 5, AOE, targets=hostile
-        ReignOfBeasts = 36937, //Lv100, instant, GCD, range 3, single-target, targets=hostile
-        NobleBlood = 36938, //Lv100, instant, GCD, range 3, single-target, targets=hostile
-        LionHeart = 36939, //Lv100, instant, GCD, range 3, single-target, targets=hostile
-
-    #endregion
-
-    #region Defensive
-        Camouflage = 16140, //Lv6, instant, 90.0s CD (group 15), range 0, single-target, targets=self
-        RoyalGuard = 16142, //Lv10, instant, 2.0s CD (group 1), range 0, single-target, targets=self
-        ReleaseRoyalGuard = 32068, //Lv10, instant, 1.0s CD (group 1), range 0, single-target, targets=self
-        Nebula = 16148, //Lv38, instant, 120.0s CD (group 21), range 0, single-target, targets=self
-        Aurora = 16151, //Lv45, instant, 60.0s CD (group 19/71), range 30, single-target, targets=self/party/alliance/friendly
-        Superbolide = 16152, //Lv50, instant, 360.0s CD (group 24), range 0, single-target, targets=self
-        HeartOfLight = 16160, //Lv64, instant, 90.0s CD (group 16), range 0, AOE 30 circle, targets=self
-        HeartOfStone = 16161, //Lv68, instant, 25.0s CD (group 3), range 30, single-target, targets=self/party
-        Trajectory = 36934, //Lv56, instant, 30s CD (group 9/70) (2? charges), range 20, single-target, targets=hostile
-        HeartOfCorundum = 25758, //Lv82, instant, 25.0s CD (group 3), range 30, single-target, targets=self/party
-        GreatNebula = 36935, //Lv92, instant, 120.0s CD, range 0, single-target, targeets=self
-    #endregion
-
-    //Limit Break
-        GunmetalSoul = 17105; //LB3, instant, range 0, AOE 50 circle, targets=self, animLock=3.860
-
-    public static class Buffs
-    {
-        public const ushort
-                BrutalShell = 1898, //applied by Brutal Shell to self
-                NoMercy = 1831, //applied by No Mercy to self
-                ReadyToRip = 1842, //applied by Gnashing Fang to self
-                SonicBreak = 1837, //applied by Sonic Break to target
-                BowShock = 1838, //applied by Bow Shock to target
-                ReadyToTear = 1843, //applied by Savage Claw to self
-                ReadyToGouge = 1844, //applied by Wicked Talon to self
-                ReadyToBlast = 2686, //applied by Burst Strike to self
-                Nebula = 1834, //applied by Nebula to self
-                Rampart = 1191, //applied by Rampart to self
-                Camouflage = 1832, //applied by Camouflage to self
-                ArmsLength = 1209, //applied by Arm's Length to self
-                HeartOfLight = 1839, //applied by Heart of Light to self
-                Aurora = 1835, //applied by Aurora to self
-                Superbolide = 1836, //applied by Superbolide to self
-                HeartOfStone = 1840, //applied by Heart of Stone to self
-                HeartOfCorundum = 2683, //applied by Heart of Corundum to self
-                ClarityOfCorundum = 2684, //applied by Heart of Corundum to self
-                CatharsisOfCorundum = 2685, //applied by Heart of Corundum to self
-                RoyalGuard = 1833, //applied by Royal Guard to self
-                GreatNebula = 3838, //applied by Nebula to self
-                ReadyToRaze = 3839, //applied by Fated Circle to self
-                ReadyToBreak = 3886, //applied by No mercy to self
-                ReadyToReign = 3840; //applied by Bloodfest to target
-    }
-    public static class Debuffs
-    {
-        public const ushort
-            BowShock = 1838, //applied by Bow Shock to target
-            SonicBreak = 1837; //applied by Sonic Break to target
-    }
     public static int MaxCartridges() => TraitLevelChecked(427) ? 3 : TraitLevelChecked(257) ? 2 : 0; //Level Check helper for Maximum Ammo
 
     #region Simple Mode - Single Target
