+++ WrathCombo/Combos/PvE/MCH/MCH.cs
@@ -13,7 +13,8 @@ internal static partial class MCH
 
         protected override uint Invoke(uint actionID)
         {
-            if (actionID is not (SplitShot or HeatedSplitShot)) return actionID;
+            if (actionID is not (SplitShot or HeatedSplitShot))
+                return actionID;
 
             if (IsEnabled(CustomComboPreset.MCH_Variant_Cure) &&
                 IsEnabled(Variant.VariantCure) &&
@@ -73,9 +74,9 @@ internal static partial class MCH
                                 return Hypercharge;
                         }
 
-                            //Queen
-                            if (UseQueen(Gauge))
-                                return OriginalHook(RookAutoturret);
+                        //Queen
+                        if (UseQueen(Gauge))
+                            return OriginalHook(RookAutoturret);
 
                         // Reassemble
                         if (Reassembled(Gauge))
@@ -154,7 +155,8 @@ internal static partial class MCH
 
         protected override uint Invoke(uint actionID)
         {
-            if (actionID is not (SplitShot or HeatedSplitShot)) return actionID;
+            if (actionID is not (SplitShot or HeatedSplitShot))
+                return actionID;
 
             if (IsEnabled(CustomComboPreset.MCH_Variant_Cure) &&
                 IsEnabled(Variant.VariantCure) &&
@@ -227,10 +229,10 @@ internal static partial class MCH
                                 return Hypercharge;
                         }
 
-                            // Queen
-                            if (IsEnabled(CustomComboPreset.MCH_ST_Adv_TurretQueen) &&
-                                UseQueen(Gauge))
-                                return OriginalHook(RookAutoturret);
+                        // Queen
+                        if (IsEnabled(CustomComboPreset.MCH_ST_Adv_TurretQueen) &&
+                            UseQueen(Gauge))
+                            return OriginalHook(RookAutoturret);
 
                         // Reassemble
                         if (IsEnabled(CustomComboPreset.MCH_ST_Adv_Reassemble) &&
@@ -301,7 +303,7 @@ internal static partial class MCH
                 if (ComboAction is SplitShot && LevelChecked(OriginalHook(SlugShot)))
                     return OriginalHook(SlugShot);
 
-                if (IsEnabled(CustomComboPreset.MCH_ST_Adv_Reassemble) && Config.MCH_ST_Reassembled [4] &&
+                if (IsEnabled(CustomComboPreset.MCH_ST_Adv_Reassemble) && Config.MCH_ST_Reassembled[4] &&
                     ComboAction == OriginalHook(SlugShot) &&
                     !LevelChecked(Drill) && !HasEffect(Buffs.Reassembled) && ActionReady(Reassemble))
                     return Reassemble;
@@ -319,7 +321,8 @@ internal static partial class MCH
 
         protected override uint Invoke(uint actionID)
         {
-            if (actionID is not (SpreadShot or Scattergun)) return actionID;
+            if (actionID is not (SpreadShot or Scattergun))
+                return actionID;
 
             if (IsEnabled(CustomComboPreset.MCH_Variant_Cure) &&
                 IsEnabled(Variant.VariantCure) &&
@@ -428,26 +431,27 @@ internal static partial class MCH
 
         protected override uint Invoke(uint actionID)
         {
-            if (actionID is not (SpreadShot or Scattergun)) return actionID;
+            if (actionID is not (SpreadShot or Scattergun))
+                return actionID;
 
             bool reassembledScattergunAoE = IsEnabled(CustomComboPreset.MCH_AoE_Adv_Reassemble) &&
-                                            Config.MCH_AoE_Reassembled [0] && HasEffect(Buffs.Reassembled);
+                                            Config.MCH_AoE_Reassembled[0] && HasEffect(Buffs.Reassembled);
 
             bool reassembledChainsawAoE =
-                (IsEnabled(CustomComboPreset.MCH_AoE_Adv_Reassemble) && Config.MCH_AoE_Reassembled [2] && HasEffect(Buffs.Reassembled)) ||
-                (IsEnabled(CustomComboPreset.MCH_AoE_Adv_Reassemble) && !Config.MCH_AoE_Reassembled [2] && !HasEffect(Buffs.Reassembled)) ||
+                (IsEnabled(CustomComboPreset.MCH_AoE_Adv_Reassemble) && Config.MCH_AoE_Reassembled[2] && HasEffect(Buffs.Reassembled)) ||
+                (IsEnabled(CustomComboPreset.MCH_AoE_Adv_Reassemble) && !Config.MCH_AoE_Reassembled[2] && !HasEffect(Buffs.Reassembled)) ||
                 (!HasEffect(Buffs.Reassembled) && GetRemainingCharges(Reassemble) <= Config.MCH_AoE_ReassemblePool) ||
                 !IsEnabled(CustomComboPreset.MCH_AoE_Adv_Reassemble);
 
             bool reassembledExcavatorAoE =
-                (IsEnabled(CustomComboPreset.MCH_AoE_Adv_Reassemble) && Config.MCH_AoE_Reassembled [3] && HasEffect(Buffs.Reassembled)) ||
-                (IsEnabled(CustomComboPreset.MCH_AoE_Adv_Reassemble) && !Config.MCH_AoE_Reassembled [3] && !HasEffect(Buffs.Reassembled)) ||
+                (IsEnabled(CustomComboPreset.MCH_AoE_Adv_Reassemble) && Config.MCH_AoE_Reassembled[3] && HasEffect(Buffs.Reassembled)) ||
+                (IsEnabled(CustomComboPreset.MCH_AoE_Adv_Reassemble) && !Config.MCH_AoE_Reassembled[3] && !HasEffect(Buffs.Reassembled)) ||
                 (!HasEffect(Buffs.Reassembled) && GetRemainingCharges(Reassemble) <= Config.MCH_AoE_ReassemblePool) ||
                 !IsEnabled(CustomComboPreset.MCH_AoE_Adv_Reassemble);
 
             bool reassembledAirAnchorAoE =
-                (IsEnabled(CustomComboPreset.MCH_AoE_Adv_Reassemble) && Config.MCH_AoE_Reassembled [1] && HasEffect(Buffs.Reassembled)) ||
-                (IsEnabled(CustomComboPreset.MCH_AoE_Adv_Reassemble) && !Config.MCH_AoE_Reassembled [1] && !HasEffect(Buffs.Reassembled)) ||
+                (IsEnabled(CustomComboPreset.MCH_AoE_Adv_Reassemble) && Config.MCH_AoE_Reassembled[1] && HasEffect(Buffs.Reassembled)) ||
+                (IsEnabled(CustomComboPreset.MCH_AoE_Adv_Reassemble) && !Config.MCH_AoE_Reassembled[1] && !HasEffect(Buffs.Reassembled)) ||
                 (!HasEffect(Buffs.Reassembled) && GetRemainingCharges(Reassemble) <= Config.MCH_AoE_ReassemblePool) ||
                 !IsEnabled(CustomComboPreset.MCH_AoE_Adv_Reassemble);
 
@@ -496,10 +500,10 @@ internal static partial class MCH
                     if (IsEnabled(CustomComboPreset.MCH_AoE_Adv_Reassemble) && !HasEffect(Buffs.Wildfire) &&
                         !HasEffect(Buffs.Reassembled) && HasCharges(Reassemble) && !JustUsed(Flamethrower, 10f) &&
                         GetRemainingCharges(Reassemble) > Config.MCH_AoE_ReassemblePool &&
-                        ((Config.MCH_AoE_Reassembled [0] && Scattergun.LevelChecked()) ||
-                         (Gauge.IsOverheated && Config.MCH_AoE_Reassembled [1] && AutoCrossbow.LevelChecked()) ||
-                         (GetCooldownRemainingTime(Chainsaw) < 1 && Config.MCH_AoE_Reassembled [2] && Chainsaw.LevelChecked()) ||
-                         (GetCooldownRemainingTime(OriginalHook(Chainsaw)) < 1 && Config.MCH_AoE_Reassembled [3] &&
+                        ((Config.MCH_AoE_Reassembled[0] && Scattergun.LevelChecked()) ||
+                         (Gauge.IsOverheated && Config.MCH_AoE_Reassembled[1] && AutoCrossbow.LevelChecked()) ||
+                         (GetCooldownRemainingTime(Chainsaw) < 1 && Config.MCH_AoE_Reassembled[2] && Chainsaw.LevelChecked()) ||
+                         (GetCooldownRemainingTime(OriginalHook(Chainsaw)) < 1 && Config.MCH_AoE_Reassembled[3] &&
                           Excavator.LevelChecked())))
                         return Reassemble;
 
@@ -591,7 +595,8 @@ internal static partial class MCH
 
         protected override uint Invoke(uint actionID)
         {
-            if (actionID is not (Heatblast or BlazingShot)) return actionID;
+            if (actionID is not (Heatblast or BlazingShot))
+                return actionID;
 
             if (IsEnabled(CustomComboPreset.MCH_Heatblast_AutoBarrel) &&
                 ActionReady(BarrelStabilizer) && !Gauge.IsOverheated)
@@ -634,7 +639,8 @@ internal static partial class MCH
 
         protected override uint Invoke(uint actionID)
         {
-            if (actionID is not AutoCrossbow) return actionID;
+            if (actionID is not AutoCrossbow)
+                return actionID;
 
             if (IsEnabled(CustomComboPreset.MCH_AutoCrossbow_AutoBarrel) &&
                 ActionReady(BarrelStabilizer) && !Gauge.IsOverheated)
@@ -673,7 +679,8 @@ internal static partial class MCH
 
         protected override uint Invoke(uint actionID)
         {
-            if (actionID is not (GaussRound or Ricochet or CheckMate or DoubleCheck)) return actionID;
+            if (actionID is not (GaussRound or Ricochet or CheckMate or DoubleCheck))
+                return actionID;
 
             if (ActionReady(OriginalHook(GaussRound)) &&
                 GetRemainingCharges(OriginalHook(GaussRound)) >=
@@ -705,7 +712,8 @@ internal static partial class MCH
 
         protected override uint Invoke(uint actionID)
         {
-            if (actionID is not (Drill or HotShot or AirAnchor or Chainsaw)) return actionID;
+            if (actionID is not (Drill or HotShot or AirAnchor or Chainsaw))
+                return actionID;
 
             if (LevelChecked(Excavator) && HasEffect(Buffs.ExcavatorReady))
                 return CalcBestAction(actionID, Excavator, Chainsaw, AirAnchor, Drill);
@@ -747,71 +755,4 @@ internal static partial class MCH
                 ? OriginalHook(11)
                 : actionID;
     }
-
-    #region ID's
-
-    public const byte JobID = 31;
-
-    public const uint
-        CleanShot = 2873,
-        HeatedCleanShot = 7413,
-        SplitShot = 2866,
-        HeatedSplitShot = 7411,
-        SlugShot = 2868,
-        HeatedSlugShot = 7412,
-        GaussRound = 2874,
-        Ricochet = 2890,
-        Reassemble = 2876,
-        Drill = 16498,
-        HotShot = 2872,
-        AirAnchor = 16500,
-        Hypercharge = 17209,
-        Heatblast = 7410,
-        SpreadShot = 2870,
-        Scattergun = 25786,
-        AutoCrossbow = 16497,
-        RookAutoturret = 2864,
-        RookOverdrive = 7415,
-        AutomatonQueen = 16501,
-        QueenOverdrive = 16502,
-        Tactician = 16889,
-        Chainsaw = 25788,
-        BioBlaster = 16499,
-        BarrelStabilizer = 7414,
-        Wildfire = 2878,
-        Dismantle = 2887,
-        Flamethrower = 7418,
-        BlazingShot = 36978,
-        DoubleCheck = 36979,
-        CheckMate = 36980,
-        Excavator = 36981,
-        FullMetalField = 36982;
-
-    public static class Buffs
-    {
-        public const ushort
-            Reassembled = 851,
-            Tactician = 1951,
-            Wildfire = 1946,
-            Overheated = 2688,
-            Flamethrower = 1205,
-            Hypercharged = 3864,
-            ExcavatorReady = 3865,
-            FullMetalMachinist = 3866;
-    }
-
-    public static class Debuffs
-    {
-        public const ushort
-            Dismantled = 860,
-            Bioblaster = 1866;
-    }
-
-    public static class Traits
-    {
-        public const ushort
-            EnhancedMultiWeapon = 605;
-    }
-
-    #endregion
 }
