+++ WrathCombo/Combos/PvE/MCH/MCH_Helper.cs
@@ -11,6 +11,73 @@ namespace WrathCombo.Combos.PvE;
 
 internal static partial class MCH
 {
+    #region ID's
+
+    public const byte JobID = 31;
+
+    public const uint
+        CleanShot = 2873,
+        HeatedCleanShot = 7413,
+        SplitShot = 2866,
+        HeatedSplitShot = 7411,
+        SlugShot = 2868,
+        HeatedSlugShot = 7412,
+        GaussRound = 2874,
+        Ricochet = 2890,
+        Reassemble = 2876,
+        Drill = 16498,
+        HotShot = 2872,
+        AirAnchor = 16500,
+        Hypercharge = 17209,
+        Heatblast = 7410,
+        SpreadShot = 2870,
+        Scattergun = 25786,
+        AutoCrossbow = 16497,
+        RookAutoturret = 2864,
+        RookOverdrive = 7415,
+        AutomatonQueen = 16501,
+        QueenOverdrive = 16502,
+        Tactician = 16889,
+        Chainsaw = 25788,
+        BioBlaster = 16499,
+        BarrelStabilizer = 7414,
+        Wildfire = 2878,
+        Dismantle = 2887,
+        Flamethrower = 7418,
+        BlazingShot = 36978,
+        DoubleCheck = 36979,
+        CheckMate = 36980,
+        Excavator = 36981,
+        FullMetalField = 36982;
+
+    public static class Buffs
+    {
+        public const ushort
+            Reassembled = 851,
+            Tactician = 1951,
+            Wildfire = 1946,
+            Overheated = 2688,
+            Flamethrower = 1205,
+            Hypercharged = 3864,
+            ExcavatorReady = 3865,
+            FullMetalMachinist = 3866;
+    }
+
+    public static class Debuffs
+    {
+        public const ushort
+            Dismantled = 860,
+            Bioblaster = 1866;
+    }
+
+    public static class Traits
+    {
+        public const ushort
+            EnhancedMultiWeapon = 605;
+    }
+
+    #endregion
+
     internal static MCHOpenerMaxLevel1 Opener1 = new();
     internal static MCHGauge Gauge = GetJobGauge<MCHGauge>();
