+++ WrathCombo/Combos/PvE/MNK/MNK_Helper.cs
@@ -10,6 +10,69 @@ namespace WrathCombo.Combos.PvE;
 
 internal static partial class MNK
 {
+    #region ID's
+
+    public const byte ClassID = 2;
+    public const byte JobID = 20;
+
+    public const uint
+        Bootshine = 53,
+        TrueStrike = 54,
+        SnapPunch = 56,
+        SteeledMeditation = 36940,
+        SteelPeak = 25761,
+        TwinSnakes = 61,
+        ArmOfTheDestroyer = 62,
+        Demolish = 66,
+        Mantra = 65,
+        DragonKick = 74,
+        Rockbreaker = 70,
+        Thunderclap = 25762,
+        HowlingFist = 25763,
+        FourPointFury = 16473,
+        PerfectBalance = 69,
+        FormShift = 4262,
+        TheForbiddenChakra = 3547,
+        MasterfulBlitz = 25764,
+        RiddleOfEarth = 7394,
+        EarthsReply = 36944,
+        RiddleOfFire = 7395,
+        Brotherhood = 7396,
+        RiddleOfWind = 25766,
+        InspiritedMeditation = 36941,
+        EnlightenedMeditation = 36943,
+        Enlightenment = 16474,
+        SixSidedStar = 16476,
+        ShadowOfTheDestroyer = 25767,
+        RisingPhoenix = 25768,
+        WindsReply = 36949,
+        ForbiddenMeditation = 36942,
+        LeapingOpo = 36945,
+        RisingRaptor = 36946,
+        PouncingCoeurl = 36947,
+        TrueNorth = 7546,
+        ElixirBurst = 36948,
+        FiresReply = 36950;
+
+    internal static class Buffs
+    {
+        public const ushort
+            TwinSnakes = 101,
+            OpoOpoForm = 107,
+            RaptorForm = 108,
+            CoeurlForm = 109,
+            PerfectBalance = 110,
+            RiddleOfFire = 1181,
+            RiddleOfWind = 2687,
+            FormlessFist = 2513,
+            TrueNorth = 1250,
+            WindsRumination = 3842,
+            FiresRumination = 3843,
+            Brotherhood = 1185;
+    }
+
+    #endregion
+
     internal static MNKGauge Gauge = GetJobGauge<MNKGauge>();
     internal static MNKOpenerLogicSL MNKOpenerSL = new();
     internal static MNKOpenerLogicLL MNKOpenerLL = new();
@@ -28,7 +91,7 @@ internal static partial class MNK
 
     internal static int CoeurlChakra => Gauge.BeastChakra.Count(x => x == BeastChakra.COEURL);
 
-    public static uint DetermineCoreAbility(uint actionId, bool useTrueNorthIfEnabled)
+    internal static uint DetermineCoreAbility(uint actionId, bool useTrueNorthIfEnabled)
     {
         if (HasEffect(Buffs.OpoOpoForm) || HasEffect(Buffs.FormlessFist))
             return Gauge.OpoOpoFury == 0 && LevelChecked(DragonKick)
@@ -70,7 +133,7 @@ internal static partial class MNK
         return actionId;
     }
 
-    public static bool UsePerfectBalance()
+    internal static bool UsePerfectBalance()
     {
         if (ActionReady(PerfectBalance) && !HasEffect(Buffs.PerfectBalance) && !HasEffect(Buffs.FormlessFist))
         {
