+++ WrathCombo/Combos/PvE/PCT/PCT.cs
@@ -8,70 +8,6 @@ namespace WrathCombo.Combos.PvE;
 
 internal partial class PCT
 {
-    public const byte JobID = 42;
-
-    public const uint
-        BlizzardinCyan = 34653,
-        StoneinYellow = 34654,
-        BlizzardIIinCyan = 34659,
-        ClawMotif = 34666,
-        ClawedMuse = 34672,
-        CometinBlack = 34663,
-        CreatureMotif = 34689,
-        FireInRed = 34650,
-        AeroInGreen = 34651,
-        WaterInBlue = 34652,
-        FireIIinRed = 34656,
-        AeroIIinGreen = 34657,
-        HammerMotif = 34668,
-        WingedMuse = 34671,
-        StrikingMuse = 34674,
-        StarryMuse = 34675,
-        HammerStamp = 34678,
-        HammerBrush = 34679,
-        PolishingHammer = 34680,
-        HolyInWhite = 34662,
-        StarrySkyMotif = 34669,
-        LandscapeMotif = 34691,
-        LivingMuse = 35347,
-        MawMotif = 34667,
-        MogoftheAges = 34676,
-        PomMotif = 34664,
-        PomMuse = 34670,
-        RainbowDrip = 34688,
-        RetributionoftheMadeen = 34677,
-        ScenicMuse = 35349,
-        Smudge = 34684,
-        StarPrism = 34681,
-        SteelMuse = 35348,
-        SubtractivePalette = 34683,
-        StoneIIinYellow = 34660,
-        ThunderIIinMagenta = 34661,
-        ThunderinMagenta = 34655,
-        WaterinBlue = 34652,
-        WeaponMotif = 34690,
-        WingMotif = 34665;
-
-    public static class Buffs
-    {
-        public const ushort
-            SubtractivePalette = 3674,
-            RainbowBright = 3679,
-            HammerTime = 3680,
-            MonochromeTones = 3691,
-            StarryMuse = 3685,
-            Hyperphantasia = 3688,
-            Inspiration = 3689,
-            SubtractiveSpectrum = 3690,
-            Starstruck = 3681;
-    }
-
-    public static class Debuffs
-    {
-
-    }
-
-
     internal class PCT_ST_SimpleMode : CustomCombo
     {
         protected internal override CustomComboPreset Preset { get; } = CustomComboPreset.PCT_ST_SimpleMode;
