+++ WrathCombo/Combos/PvE/RDM/RDM.cs
@@ -6,96 +6,15 @@ namespace WrathCombo.Combos.PvE;
 
 internal partial class RDM
 {
-    //7.0 Note
-    //Gauge information is available via RDMMana
-    public const byte JobID = 35;
-
-    public const uint
-        Verthunder = 7505,
-        Veraero = 7507,
-        Veraero2 = 16525,
-        Veraero3 = 25856,
-        Verthunder2 = 16524,
-        Verthunder3 = 25855,
-        Impact = 16526,
-        Redoublement = 7516,
-        EnchantedRedoublement = 7529,
-        Zwerchhau = 7512,
-        EnchantedZwerchhau = 7528,
-        Riposte = 7504,
-        EnchantedRiposte = 7527,
-        Scatter = 7509,
-        Verstone = 7511,
-        Verfire = 7510,
-        Vercure = 7514,
-        Jolt = 7503,
-        Jolt2 = 7524,
-        Jolt3 = 37004,
-        Verholy = 7526,
-        Verflare = 7525,
-        Fleche = 7517,
-        ContreSixte = 7519,
-        Engagement = 16527,
-        Verraise = 7523,
-        Scorch = 16530,
-        Resolution = 25858,
-        Moulinet = 7513,
-        EnchantedMoulinet = 7530,
-        EnchantedMoulinetDeux = 37002,
-        EnchantedMoulinetTrois = 37003,
-        Corpsacorps = 7506,
-        Displacement = 7515,
-        Reprise = 16529,
-        ViceOfThorns = 37005,
-        GrandImpact = 37006,
-        Prefulgence = 37007,
-
-        //Buffs
-        Acceleration = 7518,
-        Manafication = 7521,
-        Embolden = 7520,
-        MagickBarrier = 25857;
-
-    public static class Buffs
-    {
-        public const ushort
-            VerfireReady = 1234,
-            VerstoneReady = 1235,
-            Dualcast = 1249,
-            Chainspell = 2560,
-            Acceleration = 1238,
-            Embolden = 1239,
-            EmboldenOthers = 1297,
-            Manafication = 1971,
-            MagickBarrier = 2707,
-            MagickedSwordPlay = 3875,
-            ThornedFlourish = 3876,
-            GrandImpactReady = 3877,
-            PrefulugenceReady = 3878;
+    internal class RDM_VariantVerCure : CustomCombo
+    {  
+        protected internal override CustomComboPreset Preset { get; } = CustomComboPreset.RDM_Variant_Cure2;
+        
+        protected override uint Invoke(uint actionID) =>            
+            actionID is Vercure && IsEnabled(Variant.VariantCure)
+            ? Variant.VariantCure
+            : actionID;
     }
-
-    public static class Debuffs
-    {
-        // public const short placeholder = 0;
-    }
-
-    public static class Traits
-    {
-        public const uint
-            EnhancedEmbolden = 620,
-            EnhancedManaficationII = 622,
-            EnhancedManaficationIII = 622,
-            EnhancedAccelerationII = 624;
-    }
-
-        internal class RDM_VariantVerCure : CustomCombo
-        {
-            protected internal override CustomComboPreset Preset { get; } = CustomComboPreset.RDM_Variant_Cure2;
-
-            protected override uint Invoke(uint actionID) =>
-                actionID is Vercure && IsEnabled(Variant.VariantCure)
-                    ? Variant.VariantCure : actionID;
-        }
 
     internal class RDM_ST_SimpleMode : CustomCombo
     {
