+++ WrathCombo/Combos/PvE/RPR/RPR.cs
@@ -188,7 +188,8 @@ internal partial class RPR
 
         protected override uint Invoke(uint actionID)
         {
-            if (actionID is not Slice) return actionID;
+            if (actionID is not Slice)
+                return actionID;
 
             int PositionalChoice = Config.RPR_Positional;
 
@@ -743,7 +744,7 @@ internal partial class RPR
         protected internal override CustomComboPreset Preset { get; } =
             CustomComboPreset.RPR_ArcaneCirclePlentifulHarvest;
 
-        protected override uint Invoke(uint actionID) => 
+        protected override uint Invoke(uint actionID) =>
             actionID is ArcaneCircle &&
             HasEffect(Buffs.ImmortalSacrifice) && LevelChecked(PlentifulHarvest)
                 ? PlentifulHarvest
@@ -755,7 +756,7 @@ internal partial class RPR
         protected internal override CustomComboPreset Preset { get; } = CustomComboPreset.RPR_Regress;
 
         protected override uint Invoke(uint actionID) =>
-            actionID is HellsEgress or HellsIngress && 
+            actionID is HellsEgress or HellsIngress &&
             FindEffect(Buffs.Threshold)?.RemainingTime <= 9
                 ? Regress
                 : actionID;
@@ -771,15 +772,15 @@ internal partial class RPR
                 actionID is not (ShadowOfDeath or BloodStalk))
                 return actionID;
 
-            bool [] soulSowOptions = Config.RPR_SoulsowOptions;
+            bool[] soulSowOptions = Config.RPR_SoulsowOptions;
             bool soulsowReady = LevelChecked(Soulsow) && !HasEffect(Buffs.Soulsow);
 
             return (soulSowOptions.Length > 0 &&
-                    ((actionID is Harpe && soulSowOptions [0]) ||
-                     (actionID is Slice && soulSowOptions [1]) ||
-                     (actionID is SpinningScythe && soulSowOptions [2]) ||
-                     (actionID is ShadowOfDeath && soulSowOptions [3]) ||
-                     (actionID is BloodStalk && soulSowOptions [4])) && soulsowReady && !InCombat()) ||
+                    ((actionID is Harpe && soulSowOptions[0]) ||
+                     (actionID is Slice && soulSowOptions[1]) ||
+                     (actionID is SpinningScythe && soulSowOptions[2]) ||
+                     (actionID is ShadowOfDeath && soulSowOptions[3]) ||
+                     (actionID is BloodStalk && soulSowOptions[4])) && soulsowReady && !InCombat()) ||
                    (IsEnabled(CustomComboPreset.RPR_Soulsow_Combat) && actionID is Harpe && !HasBattleTarget())
                 ? Soulsow
                 : actionID;
@@ -874,90 +875,4 @@ internal partial class RPR
             }
         }
     }
-
-    #region ID's
-
-    public const byte JobID = 39;
-
-    public const uint
-
-        // Single Target
-        Slice = 24373,
-        WaxingSlice = 24374,
-        InfernalSlice = 24375,
-        ShadowOfDeath = 24378,
-        SoulSlice = 24380,
-
-        // AoE
-        SpinningScythe = 24376,
-        NightmareScythe = 24377,
-        WhorlOfDeath = 24379,
-        SoulScythe = 24381,
-
-        // Unveiled
-        Gibbet = 24382,
-        Gallows = 24383,
-        Guillotine = 24384,
-        UnveiledGibbet = 24390,
-        UnveiledGallows = 24391,
-        ExecutionersGibbet = 36970,
-        ExecutionersGallows = 36971,
-        ExecutionersGuillotine = 36972,
-
-        // Reaver
-        BloodStalk = 24389,
-        GrimSwathe = 24392,
-        Gluttony = 24393,
-
-        // Sacrifice
-        ArcaneCircle = 24405,
-        PlentifulHarvest = 24385,
-
-        // Enshroud
-        Enshroud = 24394,
-        Communio = 24398,
-        LemuresSlice = 24399,
-        LemuresScythe = 24400,
-        VoidReaping = 24395,
-        CrossReaping = 24396,
-        GrimReaping = 24397,
-        Sacrificium = 36969,
-        Perfectio = 36973,
-
-        // Miscellaneous
-        HellsIngress = 24401,
-        HellsEgress = 24402,
-        Regress = 24403,
-        Harpe = 24386,
-        Soulsow = 24387,
-        HarvestMoon = 24388;
-
-    public static class Buffs
-    {
-        public const ushort
-            SoulReaver = 2587,
-            ImmortalSacrifice = 2592,
-            ArcaneCircle = 2599,
-            EnhancedGibbet = 2588,
-            EnhancedGallows = 2589,
-            EnhancedVoidReaping = 2590,
-            EnhancedCrossReaping = 2591,
-            EnhancedHarpe = 2845,
-            Enshrouded = 2593,
-            Soulsow = 2594,
-            Threshold = 2595,
-            BloodsownCircle = 2972,
-            IdealHost = 3905,
-            Oblatio = 3857,
-            Executioner = 3858,
-            PerfectioParata = 3860;
-    }
-
-    public static class Debuffs
-    {
-        public const ushort
-            DeathsDesign = 2586;
-    }
-
-    #endregion
 }
