+++ WrathCombo/Combos/PvE/SAM/SAM_Helper.cs
@@ -12,6 +12,90 @@ namespace WrathCombo.Combos.PvE;
 
 internal static partial class SAM
 {
+    #region ID's
+
+    public const byte JobID = 34;
+
+    public const uint
+        Hakaze = 7477,
+        Yukikaze = 7480,
+        Gekko = 7481,
+        Enpi = 7486,
+        Jinpu = 7478,
+        Kasha = 7482,
+        Shifu = 7479,
+        Mangetsu = 7484,
+        Fuga = 7483,
+        Oka = 7485,
+        Higanbana = 7489,
+        TenkaGoken = 7488,
+        MidareSetsugekka = 7487,
+        Shinten = 7490,
+        Kyuten = 7491,
+        Hagakure = 7495,
+        Guren = 7496,
+        Senei = 16481,
+        MeikyoShisui = 7499,
+        Seigan = 7501,
+        ThirdEye = 7498,
+        Iaijutsu = 7867,
+        TsubameGaeshi = 16483,
+        KaeshiHiganbana = 16484,
+        Shoha = 16487,
+        Ikishoten = 16482,
+        Fuko = 25780,
+        OgiNamikiri = 25781,
+        KaeshiNamikiri = 25782,
+        Yaten = 7493,
+        Gyoten = 7492,
+        KaeshiSetsugekka = 16486,
+        TendoGoken = 36965,
+        TendoKaeshiSetsugekka = 36968,
+        Zanshin = 36964,
+        TendoSetsugekka = 36966,
+        Gyofu = 36963;
+
+    public static int NumSen(SAMGauge Gauge)
+    {
+        bool ka = gauge.Sen.HasFlag(Sen.KA);
+        bool getsu = gauge.Sen.HasFlag(Sen.GETSU);
+        bool setsu = gauge.Sen.HasFlag(Sen.SETSU);
+
+        return (ka ? 1 : 0) + (getsu ? 1 : 0) + (setsu ? 1 : 0);
+    }
+
+    public static class Buffs
+    {
+        public const ushort
+            MeikyoShisui = 1233,
+            EnhancedEnpi = 1236,
+            EyesOpen = 1252,
+            OgiNamikiriReady = 2959,
+            Fuka = 1299,
+            Fugetsu = 1298,
+            TsubameReady = 4216,
+            TendoKaeshiSetsugekkaReady = 4218,
+            KaeshiGokenReady = 3852,
+            TendoKaeshiGokenReady = 4217,
+            ZanshinReady = 3855,
+            Tendo = 3856;
+    }
+
+    public static class Debuffs
+    {
+        public const ushort
+            Higanbana = 1228;
+    }
+
+    public static class Traits
+    {
+        public const ushort
+            EnhancedHissatsu = 591,
+            EnhancedMeikyoShishui2 = 593;
+    }
+
+    #endregion
+
     internal static SAMGauge gauge = GetJobGauge<SAMGauge>();
     internal static SAMOpenerMaxLevel1 Opener1 = new();
