+++ WrathCombo/Combos/PvE/WHM/WHM.cs
@@ -10,86 +10,6 @@ namespace WrathCombo.Combos.PvE;
 
 internal partial class WHM
 {
-    public const byte ClassID = 6;
-    public const byte JobID = 24;
-
-    public const uint
-
-        // Heals
-        Cure = 120,
-        Cure2 = 135,
-        Cure3 = 131,
-        Regen = 137,
-        AfflatusSolace = 16531,
-        AfflatusRapture = 16534,
-        Raise = 125,
-        Benediction = 140,
-        AfflatusMisery = 16535,
-        Medica1 = 124,
-        Medica2 = 133,
-        Medica3 = 37010,
-        Tetragrammaton = 3570,
-        DivineBenison = 7432,
-        Aquaveil = 25861,
-        DivineCaress = 37011,
-
-        // DPS
-        Glare1 = 16533,
-        Glare3 = 25859,
-        Glare4 = 37009,
-        Stone1 = 119,
-        Stone2 = 127,
-        Stone3 = 3568,
-        Stone4 = 7431,
-        Assize = 3571,
-        Holy = 139,
-        Holy3 = 25860,
-
-        // DoT
-        Aero = 121,
-        Aero2 = 132,
-        Dia = 16532,
-
-        // Buffs
-        ThinAir = 7430,
-        PresenceOfMind = 136,
-        PlenaryIndulgence = 7433;
-
-    //Action Groups
-    internal static readonly List<uint>
-        StoneGlareList = [Stone1, Stone2, Stone3, Stone4, Glare1, Glare3];
-
-    //Debuff Pairs of Actions and Debuff
-    internal static readonly Dictionary<uint, ushort>
-        AeroList = new()
-        {
-            { Aero, Debuffs.Aero },
-            { Aero2, Debuffs.Aero2 },
-            { Dia, Debuffs.Dia }
-        };
-
-    public static class Buffs
-    {
-        public const ushort
-            Regen = 158,
-            Medica2 = 150,
-            Medica3 = 3880,
-            PresenceOfMind = 157,
-            ThinAir = 1217,
-            DivineBenison = 1218,
-            Aquaveil = 2708,
-            SacredSight = 3879,
-            DivineGrace = 3881;
-    }
-
-    public static class Debuffs
-    {
-        public const ushort
-            Aero = 143,
-            Aero2 = 144,
-            Dia = 1871;
-    }
-
     internal class WHM_SolaceMisery : CustomCombo
     {
         protected internal override CustomComboPreset Preset { get; } = CustomComboPreset.WHM_SolaceMisery;
